refactor(rome_rowan): Split API/Cursor files
Split the cursor and api files into subfiles to make it easier to navigate the code. I frequently struggled to find the right place to add an implementation. I hope, this will help a bit.

Doesn't contain any logic change.
