+++ crates/rome_rowan/src/cursor.rs
@@ -81,34 +81,35 @@
 //      when the tree is mutable.
 //    - TBD
 
-use std::iter::FusedIterator;
+mod element;
+mod node;
+mod token;
+mod trivia;
+
 use std::{
-    borrow::Cow,
     cell::Cell,
-    fmt,
-    hash::{Hash, Hasher},
-    iter,
     mem::{self, ManuallyDrop},
-    ops::Range,
     ptr,
 };
 
 use countme::Count;
+pub(crate) use trivia::{SyntaxTrivia, SyntaxTriviaPiecesIterator};
 
-use crate::green::{GreenTrivia, Slot};
-use crate::{
-    green::{Child, Children},
-    TriviaPiece,
-};
+use crate::cursor::node::Siblings;
+pub(crate) use crate::cursor::token::SyntaxToken;
 use crate::{
     green::{GreenElementRef, GreenNodeData, GreenTokenData, RawSyntaxKind},
     sll,
     utility_types::Delta,
-    Direction, GreenNode, GreenToken, NodeOrToken, SyntaxText, TextRange, TextSize, TokenAtOffset,
-    WalkEvent,
+    GreenNode, GreenToken, NodeOrToken, TextRange, TextSize,
+};
+pub(crate) use element::SyntaxElement;
+pub(crate) use node::{
+    Preorder, PreorderWithTokens, SyntaxElementChildren, SyntaxNode, SyntaxNodeChildren,
+    SyntaxSlot, SyntaxSlots,
 };
 
-enum Green {
+enum GreenElement {
     Node {
         ptr: Cell<ptr::NonNull<GreenNodeData>>,
     },
@@ -125,7 +126,7 @@ struct NodeData {
     rc: Cell<u32>,
     parent: Cell<Option<ptr::NonNull<NodeData>>>,
     slot: Cell<u32>,
-    green: Green,
+    green: GreenElement,
 
     /// Invariant: never changes after NodeData is created.
     mutable: bool,
@@ -151,58 +152,6 @@ unsafe impl sll::Elem for NodeData {
     }
 }
 
-pub(crate) type SyntaxElement = NodeOrToken<SyntaxNode, SyntaxToken>;
-
-#[derive(PartialEq, Eq, Clone, Hash)]
-pub(crate) struct SyntaxTrivia {
-    offset: TextSize,
-    token: SyntaxToken,
-    is_leading: bool,
-}
-
-pub(crate) struct SyntaxNode {
-    ptr: ptr::NonNull<NodeData>,
-}
-
-impl Clone for SyntaxNode {
-    #[inline]
-    fn clone(&self) -> Self {
-        self.data().inc_rc();
-        SyntaxNode { ptr: self.ptr }
-    }
-}
-
-impl Drop for SyntaxNode {
-    #[inline]
-    fn drop(&mut self) {
-        if self.data().dec_rc() {
-            unsafe { free(self.ptr) }
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct SyntaxToken {
-    ptr: ptr::NonNull<NodeData>,
-}
-
-impl Clone for SyntaxToken {
-    #[inline]
-    fn clone(&self) -> Self {
-        self.data().inc_rc();
-        SyntaxToken { ptr: self.ptr }
-    }
-}
-
-impl Drop for SyntaxToken {
-    #[inline]
-    fn drop(&mut self) {
-        if self.data().dec_rc() {
-            unsafe { free(self.ptr) }
-        }
-    }
-}
-
 #[inline(never)]
 unsafe fn free(mut data: ptr::NonNull<NodeData>) {
     loop {
@@ -223,10 +172,10 @@ unsafe fn free(mut data: ptr::NonNull<NodeData>) {
             }
             None => {
                 match &node.green {
-                    Green::Node { ptr } => {
+                    GreenElement::Node { ptr } => {
                         let _ = GreenNode::from_raw(ptr.get());
                     }
-                    Green::Token { ptr } => {
+                    GreenElement::Token { ptr } => {
                         let _ = GreenToken::from_raw(*ptr);
                     }
                 }
@@ -242,7 +191,7 @@ impl NodeData {
         parent: Option<SyntaxNode>,
         slot: u32,
         offset: TextSize,
-        green: Green,
+        green: GreenElement,
         mutable: bool,
     ) -> ptr::NonNull<NodeData> {
         let parent = ManuallyDrop::new(parent);
@@ -317,8 +266,8 @@ impl NodeData {
     #[inline]
     fn key(&self) -> (ptr::NonNull<()>, TextSize) {
         let ptr = match &self.green {
-            Green::Node { ptr } => ptr.get().cast(),
-            Green::Token { ptr } => ptr.cast(),
+            GreenElement::Node { ptr } => ptr.get().cast(),
+            GreenElement::Token { ptr } => ptr.cast(),
         };
         (ptr, self.offset())
     }
@@ -326,7 +275,7 @@ impl NodeData {
     #[inline]
     fn parent_node(&self) -> Option<SyntaxNode> {
         let parent = self.parent()?;
-        debug_assert!(matches!(parent.green, Green::Node { .. }));
+        debug_assert!(matches!(parent.green, GreenElement::Node { .. }));
         parent.inc_rc();
         Some(SyntaxNode {
             ptr: ptr::NonNull::from(parent),
@@ -341,8 +290,8 @@ impl NodeData {
     #[inline]
     fn green(&self) -> GreenElementRef<'_> {
         match &self.green {
-            Green::Node { ptr } => GreenElementRef::Node(unsafe { &*ptr.get().as_ptr() }),
-            Green::Token { ptr } => GreenElementRef::Token(unsafe { &*ptr.as_ref() }),
+            GreenElement::Node { ptr } => GreenElementRef::Node(unsafe { &*ptr.get().as_ptr() }),
+            GreenElement::Token { ptr } => GreenElementRef::Token(unsafe { &*ptr.as_ref() }),
         }
     }
 
@@ -350,12 +299,12 @@ impl NodeData {
     #[inline]
     fn green_siblings(&self) -> Option<Siblings> {
         match &self.parent()?.green {
-            Green::Node { ptr } => {
+            GreenElement::Node { ptr } => {
                 let parent = unsafe { &*ptr.get().as_ptr() };
 
                 Some(Siblings::new(parent, self.slot()))
             }
-            Green::Token { .. } => {
+            GreenElement::Token { .. } => {
                 debug_assert!(
                     false,
                     "A token should never be a parent of a token or node."
@@ -518,8 +467,8 @@ impl NodeData {
                 NodeOrToken::Node(green) => {
                     // Child is root, so it owns the green node. Steal it!
                     let child_green = match &child.green {
-                        Green::Node { ptr } => GreenNode::from_raw(ptr.get()).into(),
-                        Green::Token { ptr } => GreenToken::from_raw(*ptr).into(),
+                        GreenElement::Node { ptr } => GreenNode::from_raw(ptr.get()).into(),
+                        GreenElement::Token { ptr } => GreenToken::from_raw(*ptr).into(),
                     };
 
                     let green = green.insert_slot(index, Some(child_green));
@@ -533,8 +482,8 @@ impl NodeData {
         let mut node = self;
         loop {
             let old_green = match &node.green {
-                Green::Node { ptr } => ptr.replace(ptr::NonNull::from(&*new_green)),
-                Green::Token { .. } => unreachable!(),
+                GreenElement::Node { ptr } => ptr.replace(ptr::NonNull::from(&*new_green)),
+                GreenElement::Token { .. } => unreachable!(),
             };
             match node.parent() {
                 Some(parent) => match parent.green() {
@@ -554,1254 +503,3 @@ impl NodeData {
         }
     }
 }
-
-pub struct SyntaxTriviaPiecesIterator {
-    pub(crate) raw: SyntaxTrivia,
-    pub(crate) next_index: usize,
-    pub(crate) next_offset: TextSize,
-    pub(crate) end_index: usize,
-    pub(crate) end_offset: TextSize,
-}
-
-impl Iterator for SyntaxTriviaPiecesIterator {
-    type Item = (TextSize, TriviaPiece);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let trivia = self.raw.get_piece(self.next_index)?;
-        let piece = (self.next_offset, *trivia);
-
-        self.next_index += 1;
-        self.next_offset += trivia.text_len();
-
-        Some(piece)
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let len = self.end_index.saturating_sub(self.next_index);
-        (len, Some(len))
-    }
-}
-
-impl DoubleEndedIterator for SyntaxTriviaPiecesIterator {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        if self.end_index == self.next_index {
-            return None;
-        }
-
-        self.end_index -= 1;
-
-        let trivia = self.raw.get_piece(self.end_index)?;
-        self.end_offset -= trivia.text_len();
-
-        Some((self.end_offset, *trivia))
-    }
-}
-
-impl ExactSizeIterator for SyntaxTriviaPiecesIterator {}
-
-impl SyntaxTrivia {
-    pub(crate) fn text(&self) -> &str {
-        let green_token = self.token.green();
-        if self.is_leading {
-            green_token.text_leading_trivia()
-        } else {
-            green_token.text_trailing_trivia()
-        }
-    }
-
-    pub(crate) fn text_range(&self) -> TextRange {
-        let green_token = self.token.green();
-        if self.is_leading {
-            TextRange::at(self.offset, green_token.leading_trivia().text_len())
-        } else {
-            let (_, trailing_len, total_len) = green_token.leading_trailing_total_len();
-            TextRange::at(self.offset + total_len - trailing_len, trailing_len)
-        }
-    }
-
-    /// Get the number of TriviaPiece inside this trivia
-    fn len(&self) -> usize {
-        self.green_trivia().len()
-    }
-
-    /// Get the total length of text of the TriviaPieces inside this trivia
-    fn text_len(&self) -> TextSize {
-        self.green_trivia().text_len()
-    }
-
-    /// Gets index-th trivia piece when the token associated with this trivia was created.
-    /// See [SyntaxTriviaPiece].
-    pub(crate) fn get_piece(&self, index: usize) -> Option<&TriviaPiece> {
-        self.green_trivia().get_piece(index)
-    }
-
-    fn green_trivia(&self) -> &GreenTrivia {
-        match self.is_leading {
-            true => self.token.green().leading_trivia(),
-            false => self.token.green().trailing_trivia(),
-        }
-    }
-
-    /// Returns the last trivia piece element
-    pub(crate) fn last(&self) -> Option<&TriviaPiece> {
-        self.green_trivia().pieces().last()
-    }
-
-    /// Iterate over all pieces of the trivia. The iterator returns the offset
-    /// of the trivia as [TextSize] and its data as [Trivia], which contains its length.
-    /// See [SyntaxTriviaPiece].
-    pub(crate) fn pieces(&self) -> SyntaxTriviaPiecesIterator {
-        SyntaxTriviaPiecesIterator {
-            raw: self.clone(),
-            next_index: 0,
-            next_offset: self.offset,
-            end_index: self.len(),
-            end_offset: self.offset + self.text_len(),
-        }
-    }
-
-    #[inline]
-    pub(crate) fn offset(&self) -> TextSize {
-        self.offset
-    }
-}
-
-impl SyntaxNode {
-    pub(crate) fn new_root(green: GreenNode) -> SyntaxNode {
-        let green = GreenNode::into_raw(green);
-        let green = Green::Node {
-            ptr: Cell::new(green),
-        };
-        SyntaxNode {
-            ptr: NodeData::new(None, 0, 0.into(), green, false),
-        }
-    }
-
-    pub(crate) fn new_root_mut(green: GreenNode) -> SyntaxNode {
-        let green = GreenNode::into_raw(green);
-        let green = Green::Node {
-            ptr: Cell::new(green),
-        };
-        SyntaxNode {
-            ptr: NodeData::new(None, 0, 0.into(), green, true),
-        }
-    }
-
-    fn new_child(
-        green: &GreenNodeData,
-        parent: SyntaxNode,
-        slot: u32,
-        offset: TextSize,
-    ) -> SyntaxNode {
-        let mutable = parent.data().mutable;
-        let green = Green::Node {
-            ptr: Cell::new(green.into()),
-        };
-        SyntaxNode {
-            ptr: NodeData::new(Some(parent), slot, offset, green, mutable),
-        }
-    }
-
-    pub fn clone_for_update(&self) -> SyntaxNode {
-        assert!(!self.data().mutable);
-        match self.parent() {
-            Some(parent) => {
-                let parent = parent.clone_for_update();
-                SyntaxNode::new_child(self.green_ref(), parent, self.data().slot(), self.offset())
-            }
-            None => SyntaxNode::new_root_mut(self.green_ref().to_owned()),
-        }
-    }
-
-    pub fn clone_subtree(&self) -> SyntaxNode {
-        SyntaxNode::new_root(self.green().into())
-    }
-
-    #[inline]
-    fn data(&self) -> &NodeData {
-        unsafe { self.ptr.as_ref() }
-    }
-
-    #[inline]
-    pub fn kind(&self) -> RawSyntaxKind {
-        self.data().kind()
-    }
-
-    #[inline]
-    fn offset(&self) -> TextSize {
-        self.data().offset()
-    }
-
-    pub(crate) fn element_in_slot(&self, slot_index: u32) -> Option<SyntaxElement> {
-        let slot = self
-            .green_ref()
-            .slots()
-            .nth(slot_index as usize)
-            .expect("Slot index out of bounds");
-
-        slot.as_ref().map(|element| {
-            SyntaxElement::new(
-                element,
-                self.clone(),
-                slot_index,
-                self.offset() + slot.rel_offset(),
-            )
-        })
-    }
-
-    pub(crate) fn slots(&self) -> SyntaxSlots {
-        SyntaxSlots::new(self.clone())
-    }
-
-    #[inline]
-    pub fn text_range(&self) -> TextRange {
-        self.data().text_range()
-    }
-
-    pub fn text_trimmed_range(&self) -> TextRange {
-        let range = self.text_range();
-        let mut start = range.start();
-        let mut end = range.end();
-
-        // Remove all trivia from the start of the node
-        let mut token = self.first_token();
-        while let Some(t) = token.take() {
-            let (leading_len, trailing_len, total_len) = t.green().leading_trailing_total_len();
-            let token_len: u32 = (total_len - leading_len - trailing_len).into();
-            if token_len == 0 {
-                start += total_len;
-                token = t.next_token();
-            } else {
-                start += leading_len;
-            }
-        }
-
-        // Remove all trivia from the end of the node
-        let mut token = self.last_token();
-        while let Some(t) = token.take() {
-            let (leading_len, trailing_len, total_len) = t.green().leading_trailing_total_len();
-            let token_len: u32 = (total_len - leading_len - trailing_len).into();
-            if token_len == 0 {
-                end -= total_len;
-                token = t.prev_token();
-            } else {
-                end -= trailing_len;
-            }
-        }
-
-        TextRange::new(start, end.max(start))
-    }
-
-    pub fn first_leading_trivia(&self) -> Option<SyntaxTrivia> {
-        self.first_token().map(|x| x.leading_trivia())
-    }
-
-    pub fn last_trailing_trivia(&self) -> Option<SyntaxTrivia> {
-        self.last_token().map(|x| x.trailing_trivia())
-    }
-
-    #[inline]
-    pub fn index(&self) -> usize {
-        self.data().slot() as usize
-    }
-
-    #[inline]
-    pub fn text(&self) -> SyntaxText {
-        SyntaxText::new(self.clone())
-    }
-
-    #[inline]
-    pub fn text_trimmed(&self) -> SyntaxText {
-        SyntaxText::with_range(self.clone(), self.text_trimmed_range())
-    }
-
-    #[inline]
-    pub(crate) fn green(&self) -> Cow<'_, GreenNodeData> {
-        let green_ref = self.green_ref();
-        match self.data().mutable {
-            false => Cow::Borrowed(green_ref),
-            true => Cow::Owned(green_ref.to_owned()),
-        }
-    }
-    #[inline]
-    fn green_ref(&self) -> &GreenNodeData {
-        self.data().green().into_node().unwrap()
-    }
-
-    #[inline]
-    pub fn parent(&self) -> Option<SyntaxNode> {
-        self.data().parent_node()
-    }
-
-    #[inline]
-    pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode> {
-        iter::successors(Some(self.clone()), SyntaxNode::parent)
-    }
-
-    #[inline]
-    pub fn children(&self) -> SyntaxNodeChildren {
-        SyntaxNodeChildren::new(self.clone())
-    }
-
-    #[inline]
-    pub fn children_with_tokens(&self) -> SyntaxElementChildren {
-        SyntaxElementChildren::new(self.clone())
-    }
-
-    pub fn first_child(&self) -> Option<SyntaxNode> {
-        self.green_ref().children().find_map(|child| {
-            child.element().into_node().map(|green| {
-                SyntaxNode::new_child(
-                    green,
-                    self.clone(),
-                    child.slot() as u32,
-                    self.offset() + child.rel_offset(),
-                )
-            })
-        })
-    }
-
-    pub fn last_child(&self) -> Option<SyntaxNode> {
-        self.green_ref().children().rev().find_map(|child| {
-            child.element().into_node().map(|green| {
-                SyntaxNode::new_child(
-                    green,
-                    self.clone(),
-                    child.slot(),
-                    self.offset() + child.rel_offset(),
-                )
-            })
-        })
-    }
-
-    pub fn first_child_or_token(&self) -> Option<SyntaxElement> {
-        self.green_ref().children().next().map(|child| {
-            SyntaxElement::new(
-                child.element(),
-                self.clone(),
-                child.slot(),
-                self.offset() + child.rel_offset(),
-            )
-        })
-    }
-    pub fn last_child_or_token(&self) -> Option<SyntaxElement> {
-        self.green_ref().children().next_back().map(|child| {
-            SyntaxElement::new(
-                child.element(),
-                self.clone(),
-                child.slot(),
-                self.offset() + child.rel_offset(),
-            )
-        })
-    }
-
-    pub fn next_sibling(&self) -> Option<SyntaxNode> {
-        self.data().next_sibling()
-    }
-    pub fn prev_sibling(&self) -> Option<SyntaxNode> {
-        self.data().prev_sibling()
-    }
-
-    pub fn next_sibling_or_token(&self) -> Option<SyntaxElement> {
-        self.data().next_sibling_or_token()
-    }
-    pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement> {
-        self.data().prev_sibling_or_token()
-    }
-
-    pub fn first_token(&self) -> Option<SyntaxToken> {
-        self.descendants_with_tokens().find_map(|x| x.into_token())
-    }
-
-    pub fn last_token(&self) -> Option<SyntaxToken> {
-        PreorderWithTokens::reverse(self.clone())
-            .filter_map(|event| match event {
-                WalkEvent::Enter(it) => Some(it),
-                WalkEvent::Leave(_) => None,
-            })
-            .find_map(|x| x.into_token())
-    }
-
-    #[inline]
-    pub fn siblings(&self, direction: Direction) -> impl Iterator<Item = SyntaxNode> {
-        iter::successors(Some(self.clone()), move |node| match direction {
-            Direction::Next => node.next_sibling(),
-            Direction::Prev => node.prev_sibling(),
-        })
-    }
-
-    #[inline]
-    pub fn siblings_with_tokens(
-        &self,
-        direction: Direction,
-    ) -> impl Iterator<Item = SyntaxElement> {
-        let me: SyntaxElement = self.clone().into();
-        iter::successors(Some(me), move |el| match direction {
-            Direction::Next => el.next_sibling_or_token(),
-            Direction::Prev => el.prev_sibling_or_token(),
-        })
-    }
-
-    #[inline]
-    pub fn descendants(&self) -> impl Iterator<Item = SyntaxNode> {
-        self.preorder().filter_map(|event| match event {
-            WalkEvent::Enter(node) => Some(node),
-            WalkEvent::Leave(_) => None,
-        })
-    }
-
-    #[inline]
-    pub fn descendants_with_tokens(&self) -> impl Iterator<Item = SyntaxElement> {
-        self.preorder_with_tokens().filter_map(|event| match event {
-            WalkEvent::Enter(it) => Some(it),
-            WalkEvent::Leave(_) => None,
-        })
-    }
-
-    #[inline]
-    pub fn preorder(&self) -> Preorder {
-        Preorder::new(self.clone())
-    }
-
-    #[inline]
-    pub fn preorder_with_tokens(&self) -> PreorderWithTokens {
-        PreorderWithTokens::new(self.clone())
-    }
-
-    pub(crate) fn preorder_slots(&self) -> SlotsPreorder {
-        SlotsPreorder::new(self.clone())
-    }
-
-    pub fn token_at_offset(&self, offset: TextSize) -> TokenAtOffset<SyntaxToken> {
-        // TODO: this could be faster if we first drill-down to node, and only
-        // then switch to token search. We should also replace explicit
-        // recursion with a loop.
-        let range = self.text_range();
-        assert!(
-            range.start() <= offset && offset <= range.end(),
-            "Bad offset: range {:?} offset {:?}",
-            range,
-            offset
-        );
-        if range.is_empty() {
-            return TokenAtOffset::None;
-        }
-
-        let mut children = self.children_with_tokens().filter(|child| {
-            let child_range = child.text_range();
-            !child_range.is_empty()
-                && (child_range.start() <= offset && offset <= child_range.end())
-        });
-
-        let left = children.next().unwrap();
-        let right = children.next();
-        assert!(children.next().is_none());
-
-        if let Some(right) = right {
-            match (left.token_at_offset(offset), right.token_at_offset(offset)) {
-                (TokenAtOffset::Single(left), TokenAtOffset::Single(right)) => {
-                    TokenAtOffset::Between(left, right)
-                }
-                _ => unreachable!(),
-            }
-        } else {
-            left.token_at_offset(offset)
-        }
-    }
-
-    pub fn covering_element(&self, range: TextRange) -> SyntaxElement {
-        let mut res: SyntaxElement = self.clone().into();
-        loop {
-            assert!(
-                res.text_range().contains_range(range),
-                "Bad range: node range {:?}, range {:?}",
-                res.text_range(),
-                range,
-            );
-            res = match &res {
-                NodeOrToken::Token(_) => return res,
-                NodeOrToken::Node(node) => match node.child_or_token_at_range(range) {
-                    Some(it) => it,
-                    None => return res,
-                },
-            };
-        }
-    }
-
-    pub fn child_or_token_at_range(&self, range: TextRange) -> Option<SyntaxElement> {
-        let rel_range = range - self.offset();
-        self.green_ref()
-            .slot_at_range(rel_range)
-            .and_then(|(index, rel_offset, slot)| {
-                slot.as_ref().map(|green| {
-                    SyntaxElement::new(
-                        green,
-                        self.clone(),
-                        index as u32,
-                        self.offset() + rel_offset,
-                    )
-                })
-            })
-    }
-
-    pub fn splice_children(&self, to_delete: Range<usize>, to_insert: Vec<SyntaxElement>) {
-        assert!(self.data().mutable, "immutable tree: {}", self);
-        for (i, child) in self.children_with_tokens().enumerate() {
-            if to_delete.contains(&i) {
-                child.detach();
-            }
-        }
-        let mut index = to_delete.start;
-        for child in to_insert {
-            self.attach_child(index, child);
-            index += 1;
-        }
-    }
-
-    pub fn detach(&self) {
-        assert!(self.data().mutable, "immutable tree: {}", self);
-        self.data().detach()
-    }
-
-    fn attach_child(&self, index: usize, child: SyntaxElement) {
-        assert!(self.data().mutable, "immutable tree: {}", self);
-        child.detach();
-        let data = match &child {
-            NodeOrToken::Node(it) => it.data(),
-            NodeOrToken::Token(it) => it.data(),
-        };
-        self.data().attach_child(index, data)
-    }
-}
-
-impl SyntaxToken {
-    fn new(
-        green: &GreenTokenData,
-        parent: SyntaxNode,
-        index: u32,
-        offset: TextSize,
-    ) -> SyntaxToken {
-        let mutable = parent.data().mutable;
-        let green = Green::Token { ptr: green.into() };
-        SyntaxToken {
-            ptr: NodeData::new(Some(parent), index, offset, green, mutable),
-        }
-    }
-
-    #[inline]
-    fn green(&self) -> &GreenTokenData {
-        match self.data().green().as_token() {
-            Some(token) => token,
-            None => {
-                panic!(
-                    "corrupted tree: a node thinks it is a token: {:?}",
-                    self.data().green().as_node().unwrap().to_string()
-                );
-            }
-        }
-    }
-
-    #[inline]
-    fn data(&self) -> &NodeData {
-        unsafe { self.ptr.as_ref() }
-    }
-
-    #[inline]
-    pub fn kind(&self) -> RawSyntaxKind {
-        self.data().kind()
-    }
-
-    #[inline]
-    pub fn text_range(&self) -> TextRange {
-        self.data().text_range()
-    }
-
-    #[inline]
-    pub fn text_trimmed_range(&self) -> TextRange {
-        let green_token = self.green();
-        let leading_len = green_token.leading_trivia().text_len();
-        let trailing_len = green_token.trailing_trivia().text_len();
-
-        let range = self.text_range();
-        TextRange::new(range.start() + leading_len, range.end() - trailing_len)
-    }
-
-    #[inline]
-    pub fn index(&self) -> usize {
-        self.data().slot() as usize
-    }
-
-    #[inline]
-    pub fn text(&self) -> &str {
-        self.green().text()
-    }
-
-    #[inline]
-    pub fn text_trimmed(&self) -> &str {
-        self.green().text_trimmed()
-    }
-
-    #[inline]
-    pub fn parent(&self) -> Option<SyntaxNode> {
-        self.data().parent_node()
-    }
-
-    #[inline]
-    pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode> {
-        std::iter::successors(self.parent(), SyntaxNode::parent)
-    }
-
-    pub fn next_sibling_or_token(&self) -> Option<SyntaxElement> {
-        self.data().next_sibling_or_token()
-    }
-    pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement> {
-        self.data().prev_sibling_or_token()
-    }
-
-    #[inline]
-    pub fn siblings_with_tokens(
-        &self,
-        direction: Direction,
-    ) -> impl Iterator<Item = SyntaxElement> {
-        let next = move |el: &SyntaxElement| match direction {
-            Direction::Next => el.next_sibling_or_token(),
-            Direction::Prev => el.prev_sibling_or_token(),
-        };
-
-        let me: SyntaxElement = self.clone().into();
-
-        iter::successors(next(&me), next)
-    }
-
-    pub fn next_token(&self) -> Option<SyntaxToken> {
-        match self.next_sibling_or_token() {
-            Some(element) => element.first_token(),
-            None => self
-                .ancestors()
-                .find_map(|it| it.next_sibling_or_token())
-                .and_then(|element| element.first_token()),
-        }
-    }
-    pub fn prev_token(&self) -> Option<SyntaxToken> {
-        match self.prev_sibling_or_token() {
-            Some(element) => element.last_token(),
-            None => self
-                .ancestors()
-                .find_map(|it| it.prev_sibling_or_token())
-                .and_then(|element| element.last_token()),
-        }
-    }
-
-    pub fn detach(&self) {
-        assert!(self.data().mutable, "immutable tree: {}", self);
-        self.data().detach()
-    }
-
-    #[inline]
-    pub fn leading_trivia(&self) -> SyntaxTrivia {
-        SyntaxTrivia {
-            offset: self.data().offset,
-            token: self.clone(),
-            is_leading: true,
-        }
-    }
-
-    #[inline]
-    pub fn trailing_trivia(&self) -> SyntaxTrivia {
-        SyntaxTrivia {
-            offset: self.data().offset,
-            token: self.clone(),
-            is_leading: false,
-        }
-    }
-}
-
-impl SyntaxElement {
-    fn new(
-        element: GreenElementRef<'_>,
-        parent: SyntaxNode,
-        slot: u32,
-        offset: TextSize,
-    ) -> SyntaxElement {
-        match element {
-            NodeOrToken::Node(node) => {
-                SyntaxNode::new_child(node, parent, slot as u32, offset).into()
-            }
-            NodeOrToken::Token(token) => {
-                SyntaxToken::new(token, parent, slot as u32, offset).into()
-            }
-        }
-    }
-
-    #[inline]
-    pub fn text_range(&self) -> TextRange {
-        match self {
-            NodeOrToken::Node(it) => it.text_range(),
-            NodeOrToken::Token(it) => it.text_range(),
-        }
-    }
-
-    #[inline]
-    pub fn index(&self) -> usize {
-        match self {
-            NodeOrToken::Node(it) => it.index(),
-            NodeOrToken::Token(it) => it.index(),
-        }
-    }
-
-    #[inline]
-    pub fn kind(&self) -> RawSyntaxKind {
-        match self {
-            NodeOrToken::Node(it) => it.kind(),
-            NodeOrToken::Token(it) => it.kind(),
-        }
-    }
-
-    #[inline]
-    pub fn parent(&self) -> Option<SyntaxNode> {
-        match self {
-            NodeOrToken::Node(it) => it.parent(),
-            NodeOrToken::Token(it) => it.parent(),
-        }
-    }
-
-    #[inline]
-    pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode> {
-        let first = match self {
-            NodeOrToken::Node(it) => Some(it.clone()),
-            NodeOrToken::Token(it) => it.parent(),
-        };
-        iter::successors(first, SyntaxNode::parent)
-    }
-
-    pub fn first_token(&self) -> Option<SyntaxToken> {
-        match self {
-            NodeOrToken::Node(it) => it.first_token(),
-            NodeOrToken::Token(it) => Some(it.clone()),
-        }
-    }
-    pub fn last_token(&self) -> Option<SyntaxToken> {
-        match self {
-            NodeOrToken::Node(it) => it.last_token(),
-            NodeOrToken::Token(it) => Some(it.clone()),
-        }
-    }
-
-    pub fn next_sibling_or_token(&self) -> Option<SyntaxElement> {
-        match self {
-            NodeOrToken::Node(it) => it.next_sibling_or_token(),
-            NodeOrToken::Token(it) => it.next_sibling_or_token(),
-        }
-    }
-    pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement> {
-        match self {
-            NodeOrToken::Node(it) => it.prev_sibling_or_token(),
-            NodeOrToken::Token(it) => it.prev_sibling_or_token(),
-        }
-    }
-
-    fn token_at_offset(&self, offset: TextSize) -> TokenAtOffset<SyntaxToken> {
-        assert!(self.text_range().start() <= offset && offset <= self.text_range().end());
-        match self {
-            NodeOrToken::Token(token) => TokenAtOffset::Single(token.clone()),
-            NodeOrToken::Node(node) => node.token_at_offset(offset),
-        }
-    }
-
-    pub fn detach(&self) {
-        match self {
-            NodeOrToken::Node(it) => it.detach(),
-            NodeOrToken::Token(it) => it.detach(),
-        }
-    }
-}
-
-// region: impls
-
-impl fmt::Debug for SyntaxTrivia {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut f = f.debug_struct("SyntaxTrivia");
-        f.field("text_range", &self.text_range());
-        f.finish()
-    }
-}
-
-impl fmt::Display for SyntaxTrivia {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(self.text(), f)
-    }
-}
-
-// Identity semantics for hash & eq
-impl PartialEq for SyntaxNode {
-    #[inline]
-    fn eq(&self, other: &SyntaxNode) -> bool {
-        self.data().key() == other.data().key()
-    }
-}
-
-impl Eq for SyntaxNode {}
-
-impl Hash for SyntaxNode {
-    #[inline]
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.data().key().hash(state);
-    }
-}
-
-impl fmt::Debug for SyntaxNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("SyntaxNode")
-            .field("kind", &self.kind())
-            .field("text_range", &self.text_range())
-            .finish()
-    }
-}
-
-impl fmt::Display for SyntaxNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.preorder_with_tokens()
-            .filter_map(|event| match event {
-                WalkEvent::Enter(NodeOrToken::Token(token)) => Some(token),
-                _ => None,
-            })
-            .try_for_each(|it| fmt::Display::fmt(&it, f))
-    }
-}
-
-// Identity semantics for hash & eq
-impl PartialEq for SyntaxToken {
-    #[inline]
-    fn eq(&self, other: &SyntaxToken) -> bool {
-        self.data().key() == other.data().key()
-    }
-}
-
-impl Eq for SyntaxToken {}
-
-impl Hash for SyntaxToken {
-    #[inline]
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.data().key().hash(state);
-    }
-}
-
-impl fmt::Display for SyntaxToken {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(self.text(), f)
-    }
-}
-
-impl From<SyntaxNode> for SyntaxElement {
-    #[inline]
-    fn from(node: SyntaxNode) -> SyntaxElement {
-        NodeOrToken::Node(node)
-    }
-}
-
-impl From<SyntaxToken> for SyntaxElement {
-    #[inline]
-    fn from(token: SyntaxToken) -> SyntaxElement {
-        NodeOrToken::Token(token)
-    }
-}
-
-// endregion
-
-// region: iterators
-
-#[derive(Clone, Debug)]
-pub(crate) struct SyntaxNodeChildren {
-    next: Option<SyntaxNode>,
-}
-
-impl SyntaxNodeChildren {
-    fn new(parent: SyntaxNode) -> SyntaxNodeChildren {
-        SyntaxNodeChildren {
-            next: parent.first_child(),
-        }
-    }
-}
-
-impl Iterator for SyntaxNodeChildren {
-    type Item = SyntaxNode;
-    fn next(&mut self) -> Option<SyntaxNode> {
-        self.next.take().map(|next| {
-            self.next = next.next_sibling();
-            next
-        })
-    }
-}
-
-#[derive(Clone, Debug, Default)]
-pub(crate) struct SyntaxElementChildren {
-    next: Option<SyntaxElement>,
-}
-
-impl SyntaxElementChildren {
-    fn new(parent: SyntaxNode) -> SyntaxElementChildren {
-        SyntaxElementChildren {
-            next: parent.first_child_or_token(),
-        }
-    }
-}
-
-impl Iterator for SyntaxElementChildren {
-    type Item = SyntaxElement;
-    fn next(&mut self) -> Option<SyntaxElement> {
-        self.next.take().map(|next| {
-            self.next = next.next_sibling_or_token();
-            next
-        })
-    }
-}
-
-pub(crate) struct Preorder {
-    start: SyntaxNode,
-    next: Option<WalkEvent<SyntaxNode>>,
-    skip_subtree: bool,
-}
-
-impl Preorder {
-    fn new(start: SyntaxNode) -> Preorder {
-        let next = Some(WalkEvent::Enter(start.clone()));
-        Preorder {
-            start,
-            next,
-            skip_subtree: false,
-        }
-    }
-
-    pub fn skip_subtree(&mut self) {
-        self.skip_subtree = true;
-    }
-
-    #[cold]
-    fn do_skip(&mut self) {
-        self.next = self.next.take().map(|next| match next {
-            WalkEvent::Enter(first_child) => WalkEvent::Leave(first_child.parent().unwrap()),
-            WalkEvent::Leave(parent) => WalkEvent::Leave(parent),
-        })
-    }
-}
-
-impl Iterator for Preorder {
-    type Item = WalkEvent<SyntaxNode>;
-
-    fn next(&mut self) -> Option<WalkEvent<SyntaxNode>> {
-        if self.skip_subtree {
-            self.do_skip();
-            self.skip_subtree = false;
-        }
-        let next = self.next.take();
-        self.next = next.as_ref().and_then(|next| {
-            Some(match next {
-                WalkEvent::Enter(node) => match node.first_child() {
-                    Some(child) => WalkEvent::Enter(child),
-                    None => WalkEvent::Leave(node.clone()),
-                },
-                WalkEvent::Leave(node) => {
-                    if node == &self.start {
-                        return None;
-                    }
-                    match node.next_sibling() {
-                        Some(sibling) => WalkEvent::Enter(sibling),
-                        None => WalkEvent::Leave(node.parent()?),
-                    }
-                }
-            })
-        });
-        next
-    }
-}
-
-pub(crate) struct PreorderWithTokens {
-    start: SyntaxElement,
-    next: Option<WalkEvent<SyntaxElement>>,
-    skip_subtree: bool,
-    direction: Direction,
-}
-
-impl PreorderWithTokens {
-    fn new(start: SyntaxNode) -> PreorderWithTokens {
-        let next = Some(WalkEvent::Enter(start.clone().into()));
-        PreorderWithTokens {
-            start: start.into(),
-            next,
-            direction: Direction::Next,
-            skip_subtree: false,
-        }
-    }
-
-    fn reverse(start: SyntaxNode) -> PreorderWithTokens {
-        let next = Some(WalkEvent::Enter(start.clone().into()));
-        PreorderWithTokens {
-            start: start.into(),
-            next,
-            direction: Direction::Prev,
-            skip_subtree: false,
-        }
-    }
-
-    pub fn skip_subtree(&mut self) {
-        self.skip_subtree = true;
-    }
-
-    #[cold]
-    fn do_skip(&mut self) {
-        self.next = self.next.take().map(|next| match next {
-            WalkEvent::Enter(first_child) => WalkEvent::Leave(first_child.parent().unwrap().into()),
-            WalkEvent::Leave(parent) => WalkEvent::Leave(parent),
-        })
-    }
-}
-
-impl Iterator for PreorderWithTokens {
-    type Item = WalkEvent<SyntaxElement>;
-
-    fn next(&mut self) -> Option<WalkEvent<SyntaxElement>> {
-        if self.skip_subtree {
-            self.do_skip();
-            self.skip_subtree = false;
-        }
-        let next = self.next.take();
-        self.next = next.as_ref().and_then(|next| {
-            Some(match next {
-                WalkEvent::Enter(el) => match el {
-                    NodeOrToken::Node(node) => {
-                        let next = match self.direction {
-                            Direction::Next => node.first_child_or_token(),
-                            Direction::Prev => node.last_child_or_token(),
-                        };
-                        match next {
-                            Some(child) => WalkEvent::Enter(child),
-                            None => WalkEvent::Leave(node.clone().into()),
-                        }
-                    }
-                    NodeOrToken::Token(token) => WalkEvent::Leave(token.clone().into()),
-                },
-                WalkEvent::Leave(el) if el == &self.start => return None,
-                WalkEvent::Leave(el) => {
-                    let next = match self.direction {
-                        Direction::Next => el.next_sibling_or_token(),
-                        Direction::Prev => el.prev_sibling_or_token(),
-                    };
-
-                    match next {
-                        Some(sibling) => WalkEvent::Enter(sibling),
-                        None => WalkEvent::Leave(el.parent()?.into()),
-                    }
-                }
-            })
-        });
-        next
-    }
-}
-
-/// Represents a cursor to a green node slot. A slot either contains an element or is empty
-/// if the child isn't present in the source.
-#[derive(Debug, Clone)]
-pub(crate) enum SyntaxSlot {
-    Node(SyntaxNode),
-    Token(SyntaxToken),
-    Empty { parent: SyntaxNode, index: u32 },
-}
-
-impl From<SyntaxElement> for SyntaxSlot {
-    fn from(element: SyntaxElement) -> Self {
-        match element {
-            SyntaxElement::Node(node) => SyntaxSlot::Node(node),
-            SyntaxElement::Token(token) => SyntaxSlot::Token(token),
-        }
-    }
-}
-
-/// Iterator over a node's slots
-#[derive(Debug, Clone)]
-pub(crate) struct SyntaxSlots {
-    next_position: u32,
-    parent: SyntaxNode,
-}
-
-impl SyntaxSlots {
-    fn new(parent: SyntaxNode) -> Self {
-        Self {
-            next_position: 0,
-            parent,
-        }
-    }
-
-    fn map_slot(&self, slot: &Slot, slot_index: u32) -> SyntaxSlot {
-        match slot {
-            Slot::Empty { .. } => SyntaxSlot::Empty {
-                parent: self.parent.clone(),
-                index: slot_index,
-            },
-            Slot::Token { rel_offset, token } => SyntaxSlot::Token(SyntaxToken::new(
-                token,
-                self.parent.clone(),
-                slot_index,
-                self.parent.offset() + rel_offset,
-            )),
-            Slot::Node { rel_offset, node } => SyntaxSlot::Node(SyntaxNode::new_child(
-                node,
-                self.parent.clone(),
-                slot_index,
-                self.parent.offset() + rel_offset,
-            )),
-        }
-    }
-}
-
-impl<'a> Iterator for SyntaxSlots {
-    type Item = SyntaxSlot;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let mut slots = self.parent.green_ref().slots();
-        let current_position = self.next_position as usize;
-        if current_position < slots.len() {
-            let next_slot = slots.nth(current_position);
-            self.next_position += 1;
-            next_slot.map(|slot| self.map_slot(slot, current_position as u32))
-        } else {
-            None
-        }
-    }
-
-    fn last(self) -> Option<Self::Item>
-    where
-        Self: Sized,
-    {
-        let mut slots = self.parent.green_ref().slots();
-        let slots_len = slots.len() as usize;
-
-        if (self.next_position as usize) < slots_len {
-            let position = slots_len - 1;
-            slots
-                .nth(position)
-                .map(|slot| self.map_slot(slot, position as u32))
-        } else {
-            // already at/passed the end
-            None
-        }
-    }
-
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.next_position += n as u32;
-        self.next()
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.parent.green_ref().slots().size_hint()
-    }
-}
-
-impl ExactSizeIterator for SyntaxSlots {
-    fn len(&self) -> usize {
-        self.parent.green_ref().slots().len()
-    }
-}
-
-impl FusedIterator for SyntaxSlots {}
-
-/// Iterator to visit a node's slots in pre-order.
-pub(crate) struct SlotsPreorder {
-    start: SyntaxNode,
-    next: Option<WalkEvent<SyntaxSlot>>,
-}
-
-impl SlotsPreorder {
-    fn new(start: SyntaxNode) -> Self {
-        let next = Some(WalkEvent::Enter(SyntaxSlot::Node(start.clone())));
-        SlotsPreorder { start, next }
-    }
-}
-
-impl Iterator for SlotsPreorder {
-    type Item = WalkEvent<SyntaxSlot>;
-
-    fn next(&mut self) -> Option<WalkEvent<SyntaxSlot>> {
-        let next = self.next.take();
-        self.next = next.as_ref().and_then(|next| {
-            Some(match next {
-                WalkEvent::Enter(slot) => match slot {
-                    SyntaxSlot::Empty { .. } | SyntaxSlot::Token(_) => {
-                        WalkEvent::Leave(slot.clone())
-                    }
-                    SyntaxSlot::Node(node) => match node.slots().next() {
-                        None => WalkEvent::Leave(SyntaxSlot::Node(node.clone())),
-                        Some(first_slot) => WalkEvent::Enter(first_slot),
-                    },
-                },
-                WalkEvent::Leave(slot) => {
-                    let (parent, slot_index) = match slot {
-                        SyntaxSlot::Empty { parent, index } => (parent.clone(), *index as usize),
-                        SyntaxSlot::Token(token) => (token.parent()?, token.index()),
-                        SyntaxSlot::Node(node) => {
-                            if node == &self.start {
-                                return None;
-                            }
-
-                            (node.parent()?, node.index())
-                        }
-                    };
-
-                    let next_slot = parent.slots().nth(slot_index + 1);
-                    match next_slot {
-                        Some(slot) => WalkEvent::Enter(slot),
-                        None => WalkEvent::Leave(SyntaxSlot::Node(parent)),
-                    }
-                }
-            })
-        });
-        next
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct Siblings<'a> {
-    parent: &'a GreenNodeData,
-    start_slot: u32,
-}
-
-impl<'a> Siblings<'a> {
-    pub fn new(parent: &'a GreenNodeData, start_slot: u32) -> Self {
-        assert!(
-            (start_slot as usize) < parent.slots().len(),
-            "Start slot {} out of bounds {}",
-            start_slot,
-            parent.slots().len()
-        );
-
-        Self { parent, start_slot }
-    }
-
-    /// Creates an iterator over the siblings following the start node.
-    /// For example, the following siblings of the if statement's condition are
-    /// * the consequence
-    /// * potentially the else clause
-    pub fn following(&self) -> Children<'a> {
-        let mut slots = self.parent.slots().enumerate();
-
-        // Navigate to the start slot so that calling `next` returns the first following sibling
-        slots.nth(self.start_slot as usize);
-
-        Children::new(slots)
-    }
-
-    /// Creates an iterator over the siblings preceding the start node in reverse order.
-    /// For example, the preceding siblings of the if statement's condition are:
-    /// * opening parentheses: (
-    /// * if keyword: if
-    pub fn previous(&self) -> impl Iterator<Item = Child<'a>> {
-        let mut slots = self.parent.slots().enumerate();
-
-        // Navigate to the start slot from the back so that calling `next_back` (or rev().next()) returns
-        // the first slot preceding the start node
-        slots.nth_back(slots.len() - 1 - self.start_slot as usize);
-
-        Children::new(slots).rev()
-    }
-}
-
-// endregion
