+++ crates/rome_rowan/src/cursor/element.rs
@@ -1,0 +1,125 @@
+use crate::cursor::{SyntaxNode, SyntaxToken};
+use crate::green::GreenElementRef;
+use crate::{NodeOrToken, RawSyntaxKind, TokenAtOffset};
+use std::iter;
+use text_size::{TextRange, TextSize};
+
+pub(crate) type SyntaxElement = NodeOrToken<SyntaxNode, SyntaxToken>;
+
+impl SyntaxElement {
+    pub(super) fn new(
+        element: GreenElementRef<'_>,
+        parent: SyntaxNode,
+        slot: u32,
+        offset: TextSize,
+    ) -> SyntaxElement {
+        match element {
+            NodeOrToken::Node(node) => {
+                SyntaxNode::new_child(node, parent, slot as u32, offset).into()
+            }
+            NodeOrToken::Token(token) => {
+                SyntaxToken::new(token, parent, slot as u32, offset).into()
+            }
+        }
+    }
+
+    #[inline]
+    pub fn text_range(&self) -> TextRange {
+        match self {
+            NodeOrToken::Node(it) => it.text_range(),
+            NodeOrToken::Token(it) => it.text_range(),
+        }
+    }
+
+    #[inline]
+    pub fn index(&self) -> usize {
+        match self {
+            NodeOrToken::Node(it) => it.index(),
+            NodeOrToken::Token(it) => it.index(),
+        }
+    }
+
+    #[inline]
+    pub fn kind(&self) -> RawSyntaxKind {
+        match self {
+            NodeOrToken::Node(it) => it.kind(),
+            NodeOrToken::Token(it) => it.kind(),
+        }
+    }
+
+    #[inline]
+    pub fn parent(&self) -> Option<SyntaxNode> {
+        match self {
+            NodeOrToken::Node(it) => it.parent(),
+            NodeOrToken::Token(it) => it.parent(),
+        }
+    }
+
+    #[inline]
+    pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode> {
+        let first = match self {
+            NodeOrToken::Node(it) => Some(it.clone()),
+            NodeOrToken::Token(it) => it.parent(),
+        };
+        iter::successors(first, SyntaxNode::parent)
+    }
+
+    pub fn first_token(&self) -> Option<SyntaxToken> {
+        match self {
+            NodeOrToken::Node(it) => it.first_token(),
+            NodeOrToken::Token(it) => Some(it.clone()),
+        }
+    }
+    pub fn last_token(&self) -> Option<SyntaxToken> {
+        match self {
+            NodeOrToken::Node(it) => it.last_token(),
+            NodeOrToken::Token(it) => Some(it.clone()),
+        }
+    }
+
+    pub fn next_sibling_or_token(&self) -> Option<SyntaxElement> {
+        match self {
+            NodeOrToken::Node(it) => it.next_sibling_or_token(),
+            NodeOrToken::Token(it) => it.next_sibling_or_token(),
+        }
+    }
+    pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement> {
+        match self {
+            NodeOrToken::Node(it) => it.prev_sibling_or_token(),
+            NodeOrToken::Token(it) => it.prev_sibling_or_token(),
+        }
+    }
+
+    pub(super) fn token_at_offset(&self, offset: TextSize) -> TokenAtOffset<SyntaxToken> {
+        assert!(self.text_range().start() <= offset && offset <= self.text_range().end());
+        match self {
+            NodeOrToken::Token(token) => TokenAtOffset::Single(token.clone()),
+            NodeOrToken::Node(node) => node.token_at_offset(offset),
+        }
+    }
+
+    pub fn detach(&self) {
+        match self {
+            NodeOrToken::Node(it) => it.detach(),
+            NodeOrToken::Token(it) => it.detach(),
+        }
+    }
+}
+
+// region: impls
+
+impl From<SyntaxNode> for SyntaxElement {
+    #[inline]
+    fn from(node: SyntaxNode) -> SyntaxElement {
+        NodeOrToken::Node(node)
+    }
+}
+
+impl From<SyntaxToken> for SyntaxElement {
+    #[inline]
+    fn from(token: SyntaxToken) -> SyntaxElement {
+        NodeOrToken::Token(token)
+    }
+}
+
+// endregion
