+++ crates/rome_rowan/src/cursor/token.rs
@@ -1,0 +1,184 @@
+use crate::cursor::{free, GreenElement, NodeData, SyntaxElement, SyntaxNode, SyntaxTrivia};
+use crate::{Direction, GreenTokenData, RawSyntaxKind};
+use std::hash::{Hash, Hasher};
+use std::{fmt, iter, ptr};
+use text_size::{TextRange, TextSize};
+
+#[derive(Debug)]
+pub(crate) struct SyntaxToken {
+    ptr: ptr::NonNull<NodeData>,
+}
+
+impl Clone for SyntaxToken {
+    #[inline]
+    fn clone(&self) -> Self {
+        self.data().inc_rc();
+        SyntaxToken { ptr: self.ptr }
+    }
+}
+
+impl Drop for SyntaxToken {
+    #[inline]
+    fn drop(&mut self) {
+        if self.data().dec_rc() {
+            unsafe { free(self.ptr) }
+        }
+    }
+}
+
+impl SyntaxToken {
+    pub(super) fn new(
+        green: &GreenTokenData,
+        parent: SyntaxNode,
+        index: u32,
+        offset: TextSize,
+    ) -> SyntaxToken {
+        let mutable = parent.data().mutable;
+        let green = GreenElement::Token { ptr: green.into() };
+        SyntaxToken {
+            ptr: NodeData::new(Some(parent), index, offset, green, mutable),
+        }
+    }
+
+    #[inline]
+    pub(super) fn green(&self) -> &GreenTokenData {
+        match self.data().green().as_token() {
+            Some(token) => token,
+            None => {
+                panic!(
+                    "corrupted tree: a node thinks it is a token: {:?}",
+                    self.data().green().as_node().unwrap().to_string()
+                );
+            }
+        }
+    }
+
+    #[inline]
+    pub(super) fn data(&self) -> &NodeData {
+        unsafe { self.ptr.as_ref() }
+    }
+
+    #[inline]
+    pub fn kind(&self) -> RawSyntaxKind {
+        self.data().kind()
+    }
+
+    #[inline]
+    pub fn text_range(&self) -> TextRange {
+        self.data().text_range()
+    }
+
+    #[inline]
+    pub fn text_trimmed_range(&self) -> TextRange {
+        let green_token = self.green();
+        let leading_len = green_token.leading_trivia().text_len();
+        let trailing_len = green_token.trailing_trivia().text_len();
+
+        let range = self.text_range();
+        TextRange::new(range.start() + leading_len, range.end() - trailing_len)
+    }
+
+    #[inline]
+    pub fn index(&self) -> usize {
+        self.data().slot() as usize
+    }
+
+    #[inline]
+    pub fn text(&self) -> &str {
+        self.green().text()
+    }
+
+    #[inline]
+    pub fn text_trimmed(&self) -> &str {
+        self.green().text_trimmed()
+    }
+
+    #[inline]
+    pub fn parent(&self) -> Option<SyntaxNode> {
+        self.data().parent_node()
+    }
+
+    #[inline]
+    pub fn ancestors(&self) -> impl Iterator<Item = SyntaxNode> {
+        std::iter::successors(self.parent(), SyntaxNode::parent)
+    }
+
+    pub fn next_sibling_or_token(&self) -> Option<SyntaxElement> {
+        self.data().next_sibling_or_token()
+    }
+    pub fn prev_sibling_or_token(&self) -> Option<SyntaxElement> {
+        self.data().prev_sibling_or_token()
+    }
+
+    #[inline]
+    pub fn siblings_with_tokens(
+        &self,
+        direction: Direction,
+    ) -> impl Iterator<Item = SyntaxElement> {
+        let next = move |el: &SyntaxElement| match direction {
+            Direction::Next => el.next_sibling_or_token(),
+            Direction::Prev => el.prev_sibling_or_token(),
+        };
+
+        let me: SyntaxElement = self.clone().into();
+
+        iter::successors(next(&me), next)
+    }
+
+    pub fn next_token(&self) -> Option<SyntaxToken> {
+        match self.next_sibling_or_token() {
+            Some(element) => element.first_token(),
+            None => self
+                .ancestors()
+                .find_map(|it| it.next_sibling_or_token())
+                .and_then(|element| element.first_token()),
+        }
+    }
+    pub fn prev_token(&self) -> Option<SyntaxToken> {
+        match self.prev_sibling_or_token() {
+            Some(element) => element.last_token(),
+            None => self
+                .ancestors()
+                .find_map(|it| it.prev_sibling_or_token())
+                .and_then(|element| element.last_token()),
+        }
+    }
+
+    pub fn detach(&self) {
+        assert!(self.data().mutable, "immutable tree: {}", self);
+        self.data().detach()
+    }
+
+    #[inline]
+    pub fn leading_trivia(&self) -> SyntaxTrivia {
+        SyntaxTrivia::leading(self.data().offset, self.clone())
+    }
+
+    #[inline]
+    pub fn trailing_trivia(&self) -> SyntaxTrivia {
+        SyntaxTrivia::trailing(self.data().offset, self.clone())
+    }
+}
+
+// Identity semantics for hash & eq
+impl PartialEq for SyntaxToken {
+    #[inline]
+    fn eq(&self, other: &SyntaxToken) -> bool {
+        self.data().key() == other.data().key()
+    }
+}
+
+impl Eq for SyntaxToken {}
+
+impl Hash for SyntaxToken {
+    #[inline]
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.data().key().hash(state);
+    }
+}
+
+impl fmt::Display for SyntaxToken {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(self.text(), f)
+    }
+}
